+++ www/index-bit6.js
@@ -2,13 +2,15 @@ cordova.require("com.bit6.sdk.Bit6SDK");
 cordova.require("com.bit6.sdk.MyRtc");
 cordova.require("com.bit6.sdk.MyCapture");
 cordova.require("com.bit6.sdk.MySurface");
+
+var pushWrappers = cordova.require("com.bit6.sdk.PushWrappers");
 var phonertc = cordova.require("com.bit6.sdk.PhoneRTC");
 
 exports.init = function(opts) {
 
     var b6 = new bit6.Client(opts);
     // Init DeviceId and Push support
-    initPushService(b6);
+    initPushService(b6, opts.pushSupport);
     // Init native WebRTC component?
     // Check for PeerConnection instead?
     var hasWebRTC = navigator.getUserMedia || navigator.mozGetUserMedia || navigator.webkitGetUserMedia || window.getUserMedia;
@@ -31,12 +33,18 @@ exports.init = function(opts) {
             //console.log('ToSurface: ' + v + ' parent ' + v.parentNode);
             surface.onVideo(v, d, op);
         };
+
+        if (device.platform.toLowerCase() === 'android') { //Supporting only android for now
+            b6.switchCamera = function() {
+                phonertc.switchCamera();
+            }
+        }
     }
 
     return b6;
 };
 
-function initPushService(b6) {
+function initPushService(b6, customPushSupport) {
     // Short name for the platform for Bit6
     var plat = device.platform.toLowerCase();
 
@@ -69,14 +77,17 @@ function initPushService(b6) {
     }
 
     // Push support
-
-    // Helper functions for success / error responses
-    var errh = function(r) {
-        console.log('Push Error: ' + r);
-    };
-    var okh = function(r) {
-        console.log('Push Success: ' + r);
-    };
+    pushSupport = null;
+    if (customPushSupport) {
+        pushSupport = customPushSupport;
+    }
+    // TODO: Move this logic to pushWrappers since it contains both wrapper implementations?
+    // Add a factory method in the wrappers file?
+    else if (typeof PushNotification !== 'undefined' && PushNotification.init) { //PluginPush
+        pushSupport = pushWrappers.pluginPushWrapper;
+    } else if (window.plugins && window.plugins.pushNotification && window.plugins.pushNotification.register) { //legacy PushPlugin
+        pushSupport = pushWrappers.legacyPushPluginWrapper;
+    }
 
     // Helper function for submitting the push key to the server
     var sendPushkeyToServer = function(pushkey) {
@@ -96,152 +107,50 @@ function initPushService(b6) {
 
     var sendApnsPushkeyToServer = function(key) {
         //For iOS adding prefix p_/d_ to the push token for Bit6 server to use correct APNS
+        // TODO: Do we want to call isApnsProduction() just once during init?
         phonertc.isApnsProduction(function(isProd) {
             var prefix = isProd ? 'p_' : 'd_';
             sendPushkeyToServer(prefix + key);
         });
     };
 
-
-    // Got notification from GCM
-    // Note that the function has to be in global scope!
-    window.onPushGCM = function(e) {
-        console.log('Got GCM: ' + e.event);
-        switch (e.event) {
-            // Registered with GCM
-            case 'registered':
-                if (e.regid && e.regid.length > 0) {
-                    // Notify the server about this GCM regId
-                    sendPushkeyToServer(e.regid);
-                }
-                break;
-                // Got a push message
-            case 'message':
-                // App is currently running
-                if (e.foreground) {
-                    console.log('GCM msg1: ' + JSON.stringify(e.payload));
-                }
-                // App launched because the user touched a notification in the notification tray
-                else if (e.coldstart) {
-                    console.log('GCM msg2: ' + JSON.stringify(e.payload));
-                }
-                // The app was in the background
-                else {
-                    console.log('GCM msg3: ' + JSON.stringify(e.payload));
-                }
-                // For now we just feed it to Bit6 JS
-                b6._handlePushRtMessage(e.payload);
-                break;
-                // Got GCM error
-            case 'error':
-                console.log('GCM err: ' + e.msg)
-                break;
-                // Unknown GCM event
-            default:
-                console.log('GCM unknown event: ' + e.event)
-                break;
-        }
-    }
-
-    // Got notification from APN
-    // Note that the function has to be in global scope!
-    window.onPushAPN = function(e) {
-        console.log('Got APN push');
-        if (e.alert) {
-            console.log('APN alert: ' + e.alert);
-        }
-        if (e.sound) {
-            // TODO: Play sound
-        }
-        if (e.badge) {
-            window.plugins.pushNotification.setApplicationIconBadgeNumber(okh, e.badge);
-        }
-        // Feed it into JS SDK
-        b6._handlePushRtMessage(e);
-    }
-
-
     // Listen to the completion of the auth procedure.
     // At that time we will have more info about the push config
     b6.session.on('auth', function() {
-        if (window.plugins && window.plugins.pushNotification && window.plugins.pushNotification.register) { //Check if this is the legacy PushPlugin
-            registerPush_legacyPlugin();
-        } else if (PushNotification && PushNotification.init) { //PluginPush
-            registerPush_pluginPush();
+        if (!pushSupport) {
+          alert("Missing push plugin. \nPlease install one of the push plugins known to Bit6, or provide your push implementation");
+          return;
         }
+
+        var pushOpts = {
+            senderId: ''
+        };
+
+        if (plat === 'and') {
+            if (b6.session.config.gcm === undefined) {
+                alert("There will be errors as long as you don't enable GCM push notifications for this app.");
+                return;
+            }
+            pushOpts.senderId = b6.session.config.gcm.senderId;
+        }
+
+        pushSupport.register(pushOpts, onPushRegistration, onPushNotification);
     });
 
-    // Legacy and Telerik Push plugins
-    var registerPush_legacyPlugin = function() {
-            // When Bit6 auth is done, we should have Android GCM senderId
-            if (plat == 'and') {
-                // TODO: check that all the values exist, showing the alert for now
-                if (b6.session.config.gcm === undefined) {
-                    alert("There will be errors as long as you don't enable GCM push notifications for this app.");
-                }
 
-                var gcmSenderId = b6.session.config.gcm.senderId;
-                console.log('GCM senderId=' + gcmSenderId);
-                var opts = {
-                    senderID: gcmSenderId,
-                    ecb: 'onPushGCM'
-                };
-                console.log('Register GCM', opts);
-                window.plugins.pushNotification.register(okh, errh, opts);
-            }
-            // On iOS we just specify the type of pushes we want to receive
-            else if (plat == 'ios') {
-                //Hack to keep the old format of push payload, otherwise telerik/legacy plugin will fail
-                bit6.Client.version = '0.9.6'
-
-                var opts = {
-                    badge: 'true',
-                    sound: 'true',
-                    alert: 'true',
-                    ecb: 'onPushAPN'
-                };
-                console.log('Register APN', opts);
-                window.plugins.pushNotification.register(sendApnsPushkeyToServer, errh, opts);
-            }
+    function onPushNotification(data) {
+        if (!data) {
+            alert("Bit6: Push data is missing!")
+            return;
         }
+        b6._handlePushRtMessage(data);
+    }
 
-    // New PhoneGap Push plugin
-    var registerPush_pluginPush = function() {
-        if (plat == 'and') {
-            var gcmSenderId = b6.session.config.gcm.senderId;
-            console.log('GCM senderId=' + gcmSenderId);
-            var push = PushNotification.init({
-                android: {
-                    senderID: gcmSenderId,
-                    clearNotifications: 'true'
-                }
-            });
-            push.on('registration', function(data) {
-                sendPushkeyToServer(data.registrationId);
-            });
-            push.on('notification', function(data) {
-                // Feed it into JS SDK
-                b6._handlePushRtMessage(data.additionalData);
-            });
-        } else if (plat == 'ios') {
-            var push = PushNotification.init({
-                ios: {
-                    alert: 'true',
-                    badge: 'true',
-                    sound: 'true',
-                    clearBadge: 'true'
-                }
-            });
-            push.on('registration', function(data) {
-                sendApnsPushkeyToServer(data.registrationId);
-            });
-            push.on('notification', function(data) {
-                if (data.count) {
-                    push.setApplicationIconBadgeNumber(okh, errh, data.count)
-                }
-                // Feed it into JS SDK
-                b6._handlePushRtMessage(data.additionalData.data);
-            });
+    function onPushRegistration(pushToken) {
+        if (plat == 'ios') {
+            sendApnsPushkeyToServer(pushToken);
+        } else {
+            sendPushkeyToServer(pushToken);
         }
     }
 }